+++ Components/Engine/Harness/Engine_Component_harness_setup.m
@@ -1,0 +1,9 @@
+%% Set-up script for Engine component
+
+% Copyright 2022 The MathWorks, Inc.
+
+Engine_refsub_Basic_params
+
+%% Initial values
+
+initial.loadInertia_speed_rpm = 1000;