+++ ChangeLog.md
@@ -4,8 +4,8 @@
 
 - Motor Drive Unit (MDU) component is revamped with
   a new harness model and tests.
-- Motor Generator 1 component is updated to use the new MDU
-  and added tests.
+- Motor Generator 1 (MG1) and Motor Generator 2 (MG2) components
+  are updated to use the new MDU component.
 
 ## Version 1.5